Added setPinsToOuptut method and a test for LED driver

+++ src/ledBarDriver.c
@@ -2,7 +2,7 @@
 #include "ledBarDriver.h"
 #include "peripherals.h"
 
-void setLeds(uint32_t * leds, uint8_t ledsSize, uint8_t numberOfLeds){
+void setLeds(uint8_t * leds, uint8_t ledsSize, uint8_t numberOfLeds){
   uint32_t * gpioAOdr = GPIOA_ODR;
   for (uint8_t i = 0; i < ledsSize; i++) {
     if (i < numberOfLeds) {
